Update TabbedChart and RocketLeague components for improved stats display; add SetData and MatchData components for detailed match analysis
